list function cleanup
