+++ src/main.rs
@@ -1,4 +1,5 @@
 mod bookmarks;
+mod list;
 mod markdown;
 mod models;
 mod search;
@@ -9,15 +10,12 @@ mod walk;
 mod write;
 
 use crate::bookmarks::mark;
-use crate::markdown::{frontmatter, get_markdown_str};
+use crate::list::list;
 use crate::search::{create_index_and_add_documents, search_index};
 use crate::settings::SETTINGS;
-use crate::walk::{has_extension, walk_files};
+use crate::utils::{expand_tilde, slugify};
 use crate::write::{create_note, prompt};
 use clap::{ArgAction, Parser, Subcommand};
-use std::path::PathBuf;
-use utils::{expand_tilde, slugify};
-use walkdir::DirEntry;
 
 #[derive(Parser)]
 struct Cli {
@@ -50,46 +48,6 @@ enum Commands {
     Index {},
     /// Search the search index
     Search { query: String },
-}
-
-fn tag_matches(entry: &DirEntry, target_tags: &[String]) -> bool {
-    if target_tags.is_empty() {
-        return true;
-    }
-
-    let path_str = match entry.path().to_str() {
-        Some(s) => s,
-        None => return false,
-    };
-    let raw_markdown = get_markdown_str(path_str);
-    if let Some(front_matter) = frontmatter(&raw_markdown) {
-        if let Some(tags) = front_matter.tags {
-            return tags.iter().any(|tag| target_tags.contains(tag));
-        }
-    }
-    false
-}
-
-fn list(notes_dir: &PathBuf, recurse_into: bool, tags: &[String]) {
-    walk_files(
-        notes_dir,
-        recurse_into,
-        |note| has_extension(note) && tag_matches(note, tags),
-        render_file,
-    );
-}
-
-fn render_file(path_str: &str) {
-    let raw_markdown = get_markdown_str(path_str);
-    if let Some(front_matter) = frontmatter(&raw_markdown) {
-        if let Some(title) = front_matter.title {
-            println!("{}\t{}", title, path_str);
-        } else {
-            println!("{}\t{}", path_str, path_str);
-        }
-    } else {
-        println!("{}\t{}", path_str, path_str);
-    }
 }
 
 fn main() {
